refactor externalObject internals into separate file
