+++ packages/delegate/src/symbols.ts
@@ -1,3 +1,3 @@
 export const UNPATHED_ERRORS_SYMBOL = Symbol('subschemaErrors');
 export const OBJECT_SUBSCHEMA_SYMBOL = Symbol('initialSubschema');
-export const FIELD_SUBSCHEMA_MAP_SYMBOL = Symbol('subschemaMap');
+export const RESPONSE_KEY_SUBSCHEMA_MAP_SYMBOL = Symbol('subschemaMap');
