+++ packages/delegate/src/types.ts
@@ -17,7 +17,7 @@ import DataLoader from 'dataloader';
 import { ExecutionRequest, ExecutionResult, Executor } from '@graphql-tools/utils';
 
 import { Subschema } from './Subschema';
-import { OBJECT_SUBSCHEMA_SYMBOL, FIELD_SUBSCHEMA_MAP_SYMBOL, UNPATHED_ERRORS_SYMBOL } from './symbols';
+import { OBJECT_SUBSCHEMA_SYMBOL, RESPONSE_KEY_SUBSCHEMA_MAP_SYMBOL, UNPATHED_ERRORS_SYMBOL } from './symbols';
 import { TypeMap } from 'graphql/type/schema';
 
 export type SchemaTransform<TContext = Record<any, string>> = (
@@ -199,6 +199,6 @@ export interface ExternalObject<TContext = Record<string, any>> {
   __typename: string;
   key: any;
   [OBJECT_SUBSCHEMA_SYMBOL]: GraphQLSchema | SubschemaConfig<any, any, any, TContext>;
-  [FIELD_SUBSCHEMA_MAP_SYMBOL]: Record<string, GraphQLSchema | SubschemaConfig<any, any, any, TContext>>;
+  [RESPONSE_KEY_SUBSCHEMA_MAP_SYMBOL]: Record<string, GraphQLSchema | SubschemaConfig<any, any, any, TContext>>;
   [UNPATHED_ERRORS_SYMBOL]: Array<GraphQLError>;
 }
